Add simple router to web
Two routes are considered equal if their method and path are equal.
Tried to compare procs but did not find a reasonable way to assess if
two procs are the same.

The console Rake task is useful so I can load all the dependencies in
the console.

The initial web app is just returning the homepage as a 200 and the rest
as 404 for testing purposes.